Rolling back r87187...
I added the #IF KATMAI check because my build machine at the time had a version of SQL Server 2005 that I had freshly installed and not updated with the latest service packs.  The property in question was first introduced in SP2.  In pre-SP2 versions of BIDS, PC Dim Naturalizer will show a message box with an exception message.

I reviewed this with Greg and we decided to leave the code as was and add a documentation item to note that PC Dim Naturalizer will not work pre-SP2.